requirements.txt
Babel==2.10.3
celery==5.2.7
django-babel==0.6.2 
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
Django==4.1.1
python-decouple==3.6